Refactor StorageHelperAndroidTest class for database manager
